pyproject.toml
[project]
name = "Chat"
version = "0.0.1"
authors = [
    {name = "Lars Mescheder", email = "larsmescheder@gmx.net"},
]
description = "Simple CLI wrapper around OpenAI chat API"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "BSD-3"}
dependencies = [
  "PyYaml",
  "openai",
  "prompt_toolkit",
]

[project.scripts]
chat = "chat.cli:cli"

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

